[TS-331] Refactor: onboading관련 api userSlice로 옮김
