+++ src/api/user/userSlice.ts
@@ -1,6 +1,8 @@
 import { userType, basicUserDataType, habitUserDataType } from "@/types/userDataType";
 import { PayloadAction, createSlice } from "@reduxjs/toolkit";
 
+import { getIsOnboarding, postOnboarding } from "@/api/user/userThunk";
+
 const initialState: userType = {
 	nickname: "",
 	genderType: "",
@@ -14,6 +16,8 @@ const initialState: userType = {
 	behaviorUnit: "",
 	firstAlert: { noon: "오전", hour: "09", minute: "50" },
 	secondAlert: { noon: "오전", hour: "10", minute: "10" },
+	isLoading: false,
+	isOnboarding: false,
 };
 
 const userSlice = createSlice({
@@ -39,6 +43,30 @@ const userSlice = createSlice({
 			state.firstAlert = action.payload.firstAlert ?? state.firstAlert;
 			state.secondAlert = action.payload.secondAlert ?? state.secondAlert;
 		},
+	},
+	extraReducers(builder) {
+		builder
+			.addCase(getIsOnboarding.pending, (state) => {
+				state.isLoading = true;
+			})
+			.addCase(getIsOnboarding.fulfilled, (state, action) => {
+				state.isLoading = false;
+				console.log(action.payload.data.onboard);
+				state.isOnboarding = action.payload.data.onboard;
+			})
+			.addCase(getIsOnboarding.rejected, (state) => {
+				state.isLoading = false;
+			})
+			.addCase(postOnboarding.pending, (state) => {
+				state.isLoading = true;
+			})
+			.addCase(postOnboarding.fulfilled, (state) => {
+				state.isLoading = false;
+				state.isOnboarding = true;
+			})
+			.addCase(postOnboarding.rejected, (state) => {
+				state.isLoading = false;
+			});
 	},
 });
